Remove Prof. prefix from name
A member is coming in with the prefix Prof. attached to the front of
their name.

Note: At the time of this commit, the current data contains multiple
members with Dr. attached to their names. This will be dealt with in
a future PR.

(Issue: https://github.com/everypolitician/everypolitician-data/issues/25705)